== UPDATE ChangeLog ==
1.  Add/Change #172 (on using large distances to extend seeds)

2.  Add/Change #176 (Remove the hack that uses coverage and see if it still works.)

3. Changed mean-1*standardDeviation;mean+1*standardDeviation to mean-3*standardDeviation;mean+3*standardDeviation

4. There are further verifications now to check if the paths to merge really overlap.

5. Add/Change #172 (on using large distances to extend seeds)
            

Result on E. coli:

        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 Ray & 86 & 4637796 & 53927 & 94923 &  311521 &  0.9950 & 0 & 0 & 0 \\



Regressions:

        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 coli-mpich2-interleaved.sh.Ray & 117 & 4621300 & 39498 & 81231 &  176272 &  0.9890 & 0 & 4 & 4 \\

--
        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 coli-mpich2-reducer.sh.Ray & 117 & 4619407 & 39482 & 81231 &  176272 &  0.9890 & 0 & 4 & 4 \\

--
        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 coli-mpich2.sh.Ray & 118 & 4621804 & 39167 & 81231 &  176272 &  0.9890 & 0 & 4 & 4 \\

--
        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 pseudomonas-interleaved-bz2.sh.Ray & 104 & 6244567 & 60043 & 149525 &  475363 &  0.9926 & 0 & 0 & 0 \\

--
        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 pseudomonas-p-bz2-gz.sh.Ray & 109 & 6234624 & 57198 & 128529 &  475347 &  0.9917 & 0 & 0 & 0 \\

--
        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 pseudomonas-p-mpich2.sh.Ray & 109 & 6162432 & 56536 & 128529 &  475347 &  0.9804 & 0 & 0 & 0 \\

--
        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 pseudomonas-s-mpich2.sh.Ray & 251 & 6182231 & 24630 & 54457 &  312927 &  0.9839 & 0 & 0 & 0 \\

--
        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 strept-mpich2-200.sh.Ray & 102 & 2021154 & 19815 & 35057 &  127908 &  0.9865 & 0 & 0 & 0 \\

--
        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 strept-mpich2-amos.sh.Ray & 102 & 2021024 & 19813 & 35057 &  127908 &  0.9865 & 0 & 0 & 0 \\

--
        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 strept-mpich2-bz2.sh.Ray & 102 & 2021024 & 19813 & 35057 &  127908 &  0.9865 & 0 & 0 & 0 \\

--
        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 strept-mpich2-interleaved-bz2.sh.Ray & 102 & 2021154 & 19815 & 35057 &  127908 &  0.9865 & 0 & 0 & 0 \\

--
        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 strept-mpich2-interleaved.sh.Ray & 102 & 2020892 & 19812 & 35057 &  127908 &  0.9865 & 0 & 0 & 0 \\

--
        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 strept-mpich2.sh.Ray & 102 & 2021024 & 19813 & 35057 &  127908 &  0.9865 & 0 & 0 & 0 \\

--
        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 strept-mpich2-simulated-errors-reducer.sh.Ray & 209 & 2006767 & 9601 & 15544 &  50523 &  0.9700 & 0 & 2 & 0 \\

--
        %  & numberOfContigs & bases & meanSize  & n50  & max   & coverage   & misassembled & mismatches & indels
 strept-mpich2-simulated-errors.sh.Ray & 212 & 2014342 & 9501 & 15265 &  50523 &  0.9700 & 0 & 2 & 0 \\



Sébastien Boisvert

+++ code/Direction.cpp
@@ -22,13 +22,13 @@
 #include<Direction.h>
 #include<stdlib.h>
 
-void Direction::constructor(int wave,int progression){
+void Direction::constructor(uint64_t wave,int progression){
 	m_wave=wave;
 	m_progression=progression;
 	m_next=NULL;
 }
 
-int Direction::getWave(){
+uint64_t Direction::getWave(){
 	return m_wave;
 }
